Changed install scripts to reflect new apt server location

+++ setup_scripts/agx_nx/install-lor-deps-jp46.sh
@@ -1,5 +1,25 @@
 #!/bin/bash
 
+# Some checks on readiness:
+
+if which nvcc > /dev/null; then
+    echo "Found nvcc in path"
+else
+    echo "Error: nvcc not found in path: $PATH"
+    echo ""
+    echo "Add nvcc to path before running this script"
+    exit
+fi
+
+if [ ! -f /etc/apt/sources.list.d/latentai-stable.list ]; then
+    echo "Error: Add latentai apt repo before running this script."
+    echo ""
+    echo "Example:"
+    echo " wget -qO - https://public.latentai.io/add_apt_repository | sudo bash"
+    echo " sudo apt update"
+    exit
+fi
+
 # OS dependencies
 sudo apt update && sudo apt upgrade -y
 sudo apt install -y python3-pip unzip
@@ -8,22 +28,24 @@ sudo ln -s /usr/include/locale.h /usr/include/xlocale.h
 # Needed for LOR/Evaluate
 sudo apt install -y libjpeg62 libjpeg8-dev
 
+pip3 install --upgrade pip
 pip3 install Cython
 pip3 install -U --no-deps numpy==1.19.4 protobuf==3.19.6 pybind11 pkgconfig
 pip3 install Pillow==8.4.0
 
 # Installing cmake using snap
-#sudo apt remove -y --purge cmake
-#hash -r
-#sudo apt install -y snapd
-#sudo snap install cmake --classic
-#hash -r
+sudo apt remove -y --purge cmake
+hash -r
+sudo apt install -y snapd
+sudo snap install cmake --classic
+hash -r
 
-# Alternative method of upsating cmake: using apt
-wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /etc/apt/trusted.gpg.d/kitware.gpg >/dev/null
-sudo apt install -y software-properties-common
-sudo add-apt-repository 'deb https://apt.kitware.com/ubuntu/ bionic main'
-sudo apt install -y cmake
+# Alternative method of updating cmake if snap is not available (e.g. docker containers):
+#
+#wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /etc/apt/trusted.gpg.d/kitware.gpg >/dev/null
+#sudo apt install -y software-properties-common
+#sudo add-apt-repository 'deb https://apt.kitware.com/ubuntu/ bionic main'
+#sudo apt install -y cmake
 
 # Pytorch and Torchvision
 wget https://nvidia.box.com/shared/static/fjtbno0vpo676a25cgvuqc1wty0fkkg6.whl -O torch-1.10.0-cp36-cp36m-linux_aarch64.whl
@@ -53,7 +75,7 @@ then
     sudo sed -i 's/TORCH_API DLContext/TORCH_API DLDevice/1' /usr/local/lib/python3.6/dist-packages/torch/include/ATen/DLConvertor.h
 fi
 
-# Needed for 2.8
+# Needed for >= LEIP 2.8
 pip3 install typing-extensions==3.10.0.2
 pip3 install --ignore-installed PyYAML==6.0
 
@@ -63,3 +85,6 @@ pip3 install huggingface-hub==0.4.0
 pip3 install omegaconf
 pip3 install pycocotools
 pip3 install --no-deps effdet==0.2.4
+
+# Installing LRE packages now required for LOR as well as examples
+sudo apt install -y latentai-runtime-cuda latentai-runtime-dev
